requirements.txt
selenium==3.141.0
pillow==9.0.1
tqdm==4.62.3